+++ src/lessons/Lesson_16/Lesson_16.tsx
@@ -1,20 +1,7 @@
 import Counter from "components/Counter/Counter";
-import { useState } from "react"
 function Lesson_16() {
-  const [value, setValue] = useState<number>(0);
-
-  const onMinusClick = () => {
-    setValue(prevValue => prevValue - 1);
-  };
-  const onPlusClick = () => {
-    setValue(prevValue => prevValue + 1);
-  };
-
   return(
-    <Counter
-      onMinusClick={onMinusClick}
-      onPlusClick={onPlusClick}
-      countValue={value}/>
+    <Counter/>
   )
 }
